tests/wpt/tests/permissions-policy/reporting/payment-report-only.https.html

<!DOCTYPE html>
<html>
  <head>
    <script src='/resources/testharness.js'></script>
    <script src='/resources/testharnessreport.js'></script>
    <script src='/resources/testdriver.js'></script>
    <script src='/resources/testdriver-vendor.js'></script>
  </head>
  <body>
    <script>
var check_report_format = ([reports, observer]) => {
  let report = reports[0];
  assert_equals(report.type, "permissions-policy-violation");
  assert_equals(report.body.featureId, "payment");
  assert_equals(report.body.disposition, "report");
};

promise_test(async t => {
  const report = new Promise(resolve => {
    new ReportingObserver((reports, observer) => resolve([reports, observer]),
                          {types: ['permissions-policy-violation']}).observe();
  });
  try {
    await test_driver.bless();
    const request = new PaymentRequest(
      [{ supportedMethods: 'https://example.com/pay' }],
      { total: { label: 'Total', amount: { currency: 'USD', value: 0 }}},
      {});
    await request.show()
    check_report_format(await report);
  } catch (err) {
    // In case the show call was rejected, the report should be generated
    // anyway. Wait for it and check the format before failing this test.
    check_report_format(await report);
    throw err;
  }
}, "PaymentRequest report only mode");
    </script>
  </body>
</html>
